rename footer (test 1)
