+++ src/components/common/footer_1.tsx
